Migration of stdout logger to a separate package
